+++ action-core/src/main/java/com/adyen/checkout/action/core/internal/ui/ActionDelegateProvider.kt
@@ -13,6 +13,7 @@ import androidx.lifecycle.SavedStateHandle
 import com.adyen.checkout.adyen3ds2.internal.provider.Adyen3DS2ComponentProvider
 import com.adyen.checkout.await.internal.provider.AwaitComponentProvider
 import com.adyen.checkout.components.core.CheckoutConfiguration
+import com.adyen.checkout.components.core.PaymentMethodTypes
 import com.adyen.checkout.components.core.action.Action
 import com.adyen.checkout.components.core.action.AwaitAction
 import com.adyen.checkout.components.core.action.BaseThreeds2Action
@@ -20,12 +21,14 @@ import com.adyen.checkout.components.core.action.QrCodeAction
 import com.adyen.checkout.components.core.action.RedirectAction
 import com.adyen.checkout.components.core.action.SdkAction
 import com.adyen.checkout.components.core.action.VoucherAction
+import com.adyen.checkout.components.core.internal.provider.ActionComponentProvider
 import com.adyen.checkout.components.core.internal.ui.ActionDelegate
 import com.adyen.checkout.components.core.internal.ui.model.DropInOverrideParams
 import com.adyen.checkout.core.exception.CheckoutException
 import com.adyen.checkout.core.internal.util.LocaleProvider
 import com.adyen.checkout.qrcode.internal.provider.QRCodeComponentProvider
 import com.adyen.checkout.redirect.internal.provider.RedirectComponentProvider
+import com.adyen.checkout.twint.internal.provider.TwintActionComponentProvider
 import com.adyen.checkout.voucher.internal.provider.VoucherComponentProvider
 import com.adyen.checkout.wechatpay.internal.provider.WeChatPayActionComponentProvider
 
@@ -46,7 +49,8 @@ internal class ActionDelegateProvider(
             is RedirectAction -> RedirectComponentProvider(dropInOverrideParams, localeProvider)
             is BaseThreeds2Action -> Adyen3DS2ComponentProvider(dropInOverrideParams, localeProvider)
             is VoucherAction -> VoucherComponentProvider(dropInOverrideParams, localeProvider)
-            is SdkAction<*> -> WeChatPayActionComponentProvider(dropInOverrideParams, localeProvider)
+            is SdkAction<*> -> getSdkActionComponentProvider(action)
+
             else -> throw CheckoutException("Can't find delegate for action: ${action.type}")
         }
 
@@ -55,5 +59,21 @@ internal class ActionDelegateProvider(
             savedStateHandle = savedStateHandle,
             application = application,
         )
+    }
+
+    private fun getSdkActionComponentProvider(
+        action: Action,
+    ): ActionComponentProvider<*, *, *> {
+        return when (action.paymentMethodType) {
+            PaymentMethodTypes.TWINT -> TwintActionComponentProvider(dropInOverrideParams, localeProvider)
+            PaymentMethodTypes.WECHAT_PAY_SDK -> WeChatPayActionComponentProvider(
+                dropInOverrideParams,
+                localeProvider,
+            )
+
+            else -> throw CheckoutException(
+                "Can't find delegate for action: ${action.type} and type: ${action.paymentMethodType}",
+            )
+        }
     }
 }